ref: rename useCombinedReducer file to makeCombinedReducer
