+++ static/app/utils/useDispatchingReducer.spec.tsx
@@ -1,6 +1,6 @@
 import {act, renderHook, waitFor} from 'sentry-test/reactTestingLibrary';
 
-import {makeCombinedReducers} from 'sentry/utils/useCombinedReducer';
+import {makeCombinedReducers} from 'sentry/utils/makeCombinedReducer';
 import {useDispatchingReducer} from 'sentry/utils/useDispatchingReducer';
 
 describe('useDispatchingReducer', () => {
